+++ Libraries/RatioValue.h
@@ -3,6 +3,6 @@
 
 NS_CORE_START
 
-USER_DEFINED_TYPE_FLOATING( RatioValue, double )
+USER_DEFINED_TYPE_FLOATING( RatioValue, float )
 
 NS_CORE_END
