+++ SintezPPDefK/DefKNuton.h
@@ -20,14 +20,14 @@ private:
 	MatrixLine									createRightParts( const System & system );
 
 	NS_CORE InternalGearRatios					geInternalGearRatioValuesFromSystem( const System & system );
-	double										calcNorm( const MatrixLine& delta );
+	float										calcNorm( const MatrixLine& delta );
 	NS_CORE InternalGearRatios					solveNuton( const Jacobi& jacobian, System& system );
 
 	NS_CORE InternalGearRatios					findK( const NS_CORE Code& code, const NS_CORE InternalGearRatios& initialKValues, const NS_CORE Ratios& iTarget );
 
 public:
 
-	
+
 	NS_CORE InternalGearRatioArray				calculate( const NS_CORE Code& code );
 };
