+++ SintezPPDefK/Defines.h
@@ -8,12 +8,12 @@
 
 NS_ARI_START
 
-	typedef double VariableValue;
-	typedef double FunctionValue;
-	typedef std::function<FunctionValue( const VariablesSet & )> Equation;
+typedef float VariableValue;
+typedef float FunctionValue;
+typedef std::function<FunctionValue( const VariablesSet & )> Equation;
 
-	typedef double								CellType;
-	typedef std::vector<CellType>				MatrixLine;
-	typedef std::vector<MatrixLine>				MatrixTable;
+typedef float								CellType;
+typedef std::vector<CellType>				MatrixLine;
+typedef std::vector<MatrixLine>				MatrixTable;
 
 NS_ARI_END
