+++ SintezPPKinCharacteristics/KinCharacteristicsTypes.h
@@ -14,8 +14,8 @@ struct KpdValue
 };
 
 typedef std::map<NS_CORE Element, int>		Z;
-typedef std::map<NS_CORE Element, double>	M;
-typedef std::map<NS_CORE Element, double>	W;
+typedef std::map<NS_CORE Element, float>	M;
+typedef std::map<NS_CORE Element, float>	W;
 typedef std::map<NS_CORE Element, char>		N;
 typedef std::map<NS_CORE GearSetNumber, KpdValue> KpdZac;
