+++ SintezPPKinCharacteristics/SolveFunctionDiv.cpp
@@ -9,21 +9,21 @@ ari::SolveFunctionDiv_p ari::SolveFunctionDiv::create()
 	return SolveFunctionDiv_p( new SolveFunctionDiv );
 }
 
-double ari::SolveFunctionDiv::calc( const IFunction_p func, const double intervalBeg, const double intervalEnd )
+float ari::SolveFunctionDiv::calc( const IFunction_p func, const float intervalBeg, const float intervalEnd )
 {
-	const double eps = 0.00001;
+	const float eps = 0.00001f;
 
-	double a = intervalBeg;
-	double b = intervalEnd;
-	double mid;
+	float a = intervalBeg;
+	float b = intervalEnd;
+	float mid;
 
-	do 
+	do
 	{
 		mid = ( a + b ) / 2;
 
-		double fa = func->calc( a );
-		double fb = func->calc( b );
-		double fm = func->calc( mid );
+		float fa = func->calc( a );
+		float fb = func->calc( b );
+		float fm = func->calc( mid );
 
 		if ( fa*fm < 0 )
 			b = mid;
@@ -38,7 +38,7 @@ double ari::SolveFunctionDiv::calc( const IFunction_p func, const double interva
 			NS_CORE Log::warning( true, "Cant solve", NS_CORE Log::NON_CRITICAL, HERE );
 			return 0;
 		}
-	} while ( abs(func->calc( ( a + b ) / 2 )) > eps );
+	} while ( abs( func->calc( ( a + b ) / 2 ) ) > eps );
 
 	return ( a + b ) / 2;
 }
